fix: window value should be int, not float

+++ clearvoice/utils/decode.py
@@ -67,7 +67,7 @@ def decode_one_audio_mossformer2_ss_16k(model, device, inputs, args):
     """
     out = []  # Initialize the list to store outputs
     decode_do_segment = False  # Flag to determine if segmentation is needed
-    window = args.sampling_rate * args.decode_window  # Decoding window length
+    window = int(args.sampling_rate * args.decode_window)  # Decoding window length
     stride = int(window * 0.75)  # Decoding stride if segmentation is used
     b, t = inputs.shape  # Get batch size and input length
 
@@ -142,7 +142,7 @@ def decode_one_audio_frcrn_se_16k(model, device, inputs, args):
     """
     decode_do_segment = False  # Flag to determine if segmentation is needed
 
-    window = args.sampling_rate * args.decode_window  # Decoding window length
+    window = int(args.sampling_rate * args.decode_window)  # Decoding window length
     stride = int(window * 0.75)  # Decoding stride for segmenting the input
     b, t = inputs.shape  # Get batch size (b) and input length (t)
 
@@ -210,7 +210,7 @@ def decode_one_audio_mossformergan_se_16k(model, device, inputs, args):
         numpy.ndarray: The decoded audio output, which has been enhanced by the model.
     """
     decode_do_segment = False  # Flag to determine if segmentation is needed
-    window = args.sampling_rate * args.decode_window  # Decoding window length
+    window = int(args.sampling_rate * args.decode_window)  # Decoding window length
     stride = int(window * 0.75)  # Decoding stride for segmenting the input
     b, t = inputs.shape  # Get batch size (b) and input length (t)
